+++ Chips/Chips.jsx
@@ -7,7 +7,7 @@ class ChipSet extends MaterialComponent {
   constructor() {
     super();
     this.componentName = 'chip-set';
-    this._mdcProps = ['choice', 'filter'];
+    this._mdcProps = ['choice', 'filter', 'input'];
   }
 
   componentDidMount() {
@@ -33,10 +33,7 @@ class Chip extends MaterialComponent {
   constructor() {
     super();
     this.componentName = 'chip';
-  }
-
-  componentWillUnmount() {
-    this.MDComponent.destroy && this.MDComponent.destroy();
+    this._mdcProps = ['selected'];
   }
 
   materialDom(allprops) {
@@ -58,7 +55,16 @@ class ChipIcon extends Icon {
   }
   materialDom(allprops) {
     const {children, ...props} = allprops;
-    return <Icon {...props}>{children}</Icon>;
+    const otherprops = {};
+    if (props.trailing) {
+      otherprops['tabindex'] = 0;
+      otherprops['role'] = 'button';
+    }
+    return (
+      <Icon {...props} {...otherprops}>
+        {children}
+      </Icon>
+    );
   }
 }